pipeline/requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorlog==6.10.1
confection==0.1.5
cymem==2.0.11
datasets==4.3.0
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docling==2.60.0
docling-core==2.50.0
docling-ibm-models==3.10.2
docling-parse==4.7.0
dspy==3.0.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.1
Faker==37.12.0
fastapi==0.121.1
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
gepa==0.0.7
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json_repair==0.52.4
jsonlines==4.0.0
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kg-gen==0.4.0
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.78.1
litellm==1.79.1
lxml==6.0.2
magicattr==0.1.6
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
numpy==2.2.6
ocrmac==1.0.0
omegaconf==2.3.0
openai==2.7.0
opencv-python==4.12.0.88
openpyxl==3.1.5
optuna==4.5.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
polyfactory==2.22.3
pre_commit==4.3.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Vision==12.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.4.3
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
ty==0.0.1a25
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
wrapt==2.0.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
